requirements.txt
aiohttp==3.9.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
Brotli==1.1.0